Refactor monitorng metric recipe for docs usage
